fixes to test
Former-commit-id: a08f801c8404ea9d4e2b09b3ed5c04ab7c6d6337

+++ tests/data/motifscan/test_motifscan.py
@@ -63,7 +63,7 @@ def test_scan():
         scores, positions, strands = chd.data.motifscan.motifscan.scan(onehot, pwm, cutoff=1.5)
 
         assert np.allclose(scores, test["expected_scores"]), (test["sequence"], scores)
-        assert np.allclose(positions, test["expected_positions"].astype(np.int))
+        assert np.allclose(positions, test["expected_positions"].astype(positions.dtype))
         assert np.allclose(strands, test["expected_strands"].astype(np.int8))
